requirements.txt
opencv-python==4.9.0.80
pytesseract==0.3.10
pokemontcgsdk==3.5.0
Pillow==10.2.0
requests==2.31.0
python-dotenv==1.0.1
